+++ frontend/src/components/StableCoin/Holders.vue
@@ -1,6 +1,6 @@
 <template>
 	<div>
-		<div v-if="holderLoading" class="w-full h-36 text-center">
+		<div v-if="pltHolderLoading" class="w-full h-36 text-center">
 			<BWCubeLogoIcon class="w-10 h-10 animate-ping mt-8" />
 		</div>
 		<div v-else>
@@ -9,45 +9,36 @@
 					<Filter
 						v-model="lastNTransactions"
 						class="mb-4"
-						:data="transactionFilterOptions"
+						:data="holderFilterOptions"
 					/>
 					<Table>
 						<TableHead>
 							<TableRow>
-								<TableTh width="25%">Account</TableTh>
-								<TableTh width="25%">Quantity</TableTh>
-								<TableTh width="25%">Percentage</TableTh>
-								<TableTh width="25%">Value</TableTh>
+								<TableTh width="33%">Account</TableTh>
+								<TableTh width="33%">Quantity</TableTh>
+								<TableTh width="1%">Percentage</TableTh>
 							</TableRow>
 						</TableHead>
 						<TableBody>
-							<TableRow
-								v-for="(coin, index) in dataPerStablecoin?.stablecoin?.holdings"
-								:key="index"
-							>
+							<TableRow v-for="(coin, index) in pagedData" :key="index">
 								<TableTd>
-									<AccountLink :address="coin.address" />
+									<AccountLink :address="coin.accountAddress.asString" />
 								</TableTd>
 								<TableTd>
-									<Amount
-										:amount="
-											coin.quantity !== undefined ? Number(coin.quantity) : 0
-										"
+									<PltAmount
+										:value="coin.amount.value"
+										:decimals="Number(coin.amount.decimals)"
 									/>
 								</TableTd>
-								<TableTd> {{ coin.percentage?.toFixed(2) }}% </TableTd>
 								<TableTd>
-									<Tooltip
-										:text="
-											String(
-												(coin.quantity ?? 0) *
-													(dataPerStablecoin?.stablecoin?.valueInDollar ?? 0)
+									<PltAmountPercentage
+										:value="
+											calculatePercentageforBigInt(
+												BigInt(coin.amount.value),
+												totalSupply
 											)
 										"
-										text-class="text-theme-body"
-									>
-										${{ numberFormatter(coin?.quantity) }}
-									</Tooltip>
+									/>
 								</TableTd>
 							</TableRow>
 						</TableBody>
@@ -58,38 +49,63 @@
 	</div>
 </template>
 <script lang="ts" setup>
-import { numberFormatter } from '~/utils/format'
-import { useStableCoinDashboardList } from '~/queries/useStableCoinDashboardList'
 import FtbCarousel from '~/components/molecules/FtbCarousel.vue'
 import CarouselSlide from '~/components/molecules/CarouselSlide.vue'
 import BWCubeLogoIcon from '~/components/icons/BWCubeLogoIcon.vue'
 import Filter from '~/components/StableCoin/Filter.vue'
+import { usePltTokenHolderQuery } from '~/queries/usePltTokenHolderQuery'
+import type { PltAccountAmount } from '~/types/generated'
+import { usePagedData } from '~/composables/usePagedData'
+
 import {
+	holderFilterOptions,
 	TransactionFilterOption,
-	transactionFilterOptions,
 } from '~/types/stable-coin'
+import { ref, watch } from 'vue'
 
 // Define Props
 const props = defineProps<{
-	coinId?: string
+	coinId: string
+	totalSupply: bigint
 }>()
 
 // Loading state
-const isLoading = ref(true)
 const lastNTransactions = ref(TransactionFilterOption.Top20)
-const limit = ref(20)
 
-const coinId = props.coinId?.toUpperCase() ?? 'USDC'
+const coinId = props.coinId
+const totalSupply = props.totalSupply
 
-// Fetch Data
+const { pagedData, addPagedData } = usePagedData<PltAccountAmount>(
+	[],
+	lastNTransactions.value,
+	lastNTransactions.value
+)
 
-const { data: dataPerStablecoin, fetching: holderLoading } =
-	useStableCoinDashboardList(coinId, limit, lastNTransactions)
+const queryFirst = ref(lastNTransactions.value)
 
-// Watch for data updates
-watch(dataPerStablecoin, newData => {
-	if (newData) {
-		isLoading.value = false
+watch(
+	() => lastNTransactions.value,
+	newValue => {
+		queryFirst.value = newValue
+		pagedData.value = []
 	}
-})
+)
+
+const { data: pltHolderData, fetching: pltHolderLoading } =
+	usePltTokenHolderQuery(coinId, {
+		first: queryFirst,
+	})
+
+watch(
+	() => pltHolderData.value,
+	value => {
+		if (value?.pltAccountsByTokenId) {
+			addPagedData(
+				value.pltAccountsByTokenId.nodes || [],
+				value.pltAccountsByTokenId.pageInfo
+			)
+		}
+	},
+	{ immediate: true }
+)
 </script>