+++ frontend/src/components/molecules/ChartCards/StableCoinSupplyBarChart.vue
@@ -11,11 +11,11 @@
 	</div>
 </template>
 <script lang="ts" setup>
-import type { Plttoken } from '~/types/generated'
+import type { PltToken } from '~/types/generated'
 import DistributionChartBar from '../DistributionChartBar.vue'
 
 type Props = {
-	stableCoinsData?: Plttoken[]
+	stableCoinsData?: PltToken[]
 	isLoading?: boolean
 }