+++ frontend/src/pages/protocol-token/index.vue
@@ -19,7 +19,7 @@
 						<template #title>Total Token Supply</template>
 						<template #value>
 							<p class="font-bold text-2xl mt-2">
-								<PLtAmount
+								<PltAmount
 									:value="
 										pltTokenDataRef
 											.reduce(
@@ -192,7 +192,7 @@
 								event.tokenEvent.__typename == 'TokenTransferEvent'
 							"
 						>
-							<PLtAmount
+							<PltAmount
 								:value="event.tokenEvent.amount.value"
 								:decimals="Number(event.tokenEvent.amount.decimals)"
 							/>
@@ -221,7 +221,7 @@ import {
 } from '~/queries/usePltTokenQuery'
 import { usePagedData } from '~/composables/usePagedData'
 import { usePltEventsQuery } from '~/queries/usePltEventsQuery'
-import type { Pltevent } from '~/types/generated'
+import type { PltEvent } from '~/types/generated'
 import { useDateNow } from '~/composables/useDateNow'
 import HolderByStableCoin from '~/components/molecules/ChartCards/HolderByStableCoin.vue'
 import KeyValueChartCard from '~/components/molecules/KeyValueChartCard.vue'
@@ -235,7 +235,7 @@ definePageMeta({
 const pageSize = 10
 const maxPageSize = 20
 const { pagedData, first, last, after, before, addPagedData, loadMore } =
-	usePagedData<Pltevent>([], pageSize, maxPageSize)
+	usePagedData<PltEvent>([], pageSize, maxPageSize)
 
 const { data: pltTokenData, loading: pltTokenLoading } = usePltTokenQuery()